# traffic-safety-openapi/public/api/docs/openapi.json
openapi: 3.0.0
info:
  title: 'Traffic Safety Platform API'
  description: 'API documentation for the Traffic Safety decision management system.'
  version: 1.0.0
servers:
  -
    url: /
    description: 'Relative base URL'
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: 'Authenticate a user and return JWT tokens.'
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              type: object
      responses:
        '200':
          description: 'Authentication successful.'
          content:
            application/json:
              schema:
                required:
                  - token
                  - tokenType
                  - expiresIn
                  - refreshToken
                  - refreshTokenExpiresAt
                  - user
                properties:
                  token: { type: string }
                  tokenType: { type: string, example: Bearer }
                  expiresIn: { type: integer, example: 3600 }
                  refreshToken: { type: string }
                  refreshTokenExpiresAt: { type: string, format: date-time }
                  user: { $ref: '#/components/schemas/User' }
                type: object
        '401':
          description: 'Invalid credentials.'
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: 'Refresh an access token using a refresh token.'
      operationId: authRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
              type: object
      responses:
        '200':
          description: 'New tokens issued.'
          content:
            application/json:
              schema:
                required:
                  - token
                  - tokenType
                  - expiresIn
                  - refreshToken
                  - refreshTokenExpiresAt
                  - user
                properties:
                  token: { type: string }
                  tokenType: { type: string }
                  expiresIn: { type: integer }
                  refreshToken: { type: string }
                  refreshTokenExpiresAt: { type: string, format: date-time }
                  user: { $ref: '#/components/schemas/User' }
                type: object
        '401':
          description: 'Refresh token invalid or expired.'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: 'Invalidate tokens and terminate the current session.'
      operationId: authLogout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              properties:
                invalidateAll:
                  description: 'Revoke all tokens for the user.'
                  type: boolean
                refreshToken:
                  type: string
                  nullable: true
              type: object
      responses:
        '204':
          description: 'Logout successful.'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    User:
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isActive
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
        role:
          type: string
          example: analyst
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...